changed int to float

+++ main.c
@@ -3,10 +3,10 @@
 
 void main(){
 //	variable declaration
-	int input,i;
+	float input,i;
 //	read data from user
     printf("enter any number:");
-    scanf("%d",&input);
+    scanf("%f",&input);
 //    logic
     for(i=1;i<=10;i++)
       printf("\n%dx%d=%d",input,i,input*i);
